[YS-479] fix: 모집 대상 popover 내부에서 저장하지 않을 경우 기존 필터링 정보 유지
